기능별 패키지 생성 및 기존 파일 분배
